move some API into profile package
