+++ hugegraph-api/src/main/java/com/baidu/hugegraph/api/job/GremlinAPI.java
@@ -50,7 +50,7 @@ import com.baidu.hugegraph.backend.query.Query;
 import com.baidu.hugegraph.core.GraphManager;
 import com.baidu.hugegraph.job.Job;
 import com.baidu.hugegraph.job.JobBuilder;
-import com.baidu.hugegraph.metric.MetricsUtil;
+import com.baidu.hugegraph.metrics.MetricsUtil;
 import com.baidu.hugegraph.server.RestServer;
 import com.baidu.hugegraph.traversal.optimize.HugeScriptTraversal;
 import com.baidu.hugegraph.util.E;
