+++ hugegraph-api/src/main/java/com/baidu/hugegraph/api/metrics/MetricsAPI.java
@@ -38,9 +38,9 @@ import com.baidu.hugegraph.api.API;
 import com.baidu.hugegraph.backend.store.BackendMetrics;
 import com.baidu.hugegraph.backend.tx.GraphTransaction;
 import com.baidu.hugegraph.core.GraphManager;
-import com.baidu.hugegraph.metric.MetricsModule;
-import com.baidu.hugegraph.metric.ServerReporter;
-import com.baidu.hugegraph.metric.SystemMetrics;
+import com.baidu.hugegraph.metrics.MetricsModule;
+import com.baidu.hugegraph.metrics.ServerReporter;
+import com.baidu.hugegraph.metrics.SystemMetrics;
 import com.baidu.hugegraph.util.InsertionOrderUtil;
 import com.baidu.hugegraph.util.JsonUtil;
 import com.baidu.hugegraph.util.Log;
