+++ hugegraph-api/src/main/java/com/baidu/hugegraph/core/GraphManager.java
@@ -44,8 +44,8 @@ import com.baidu.hugegraph.backend.store.BackendStoreSystemInfo;
 import com.baidu.hugegraph.config.HugeConfig;
 import com.baidu.hugegraph.config.ServerOptions;
 import com.baidu.hugegraph.exception.NotSupportException;
-import com.baidu.hugegraph.metric.MetricsUtil;
-import com.baidu.hugegraph.metric.ServerReporter;
+import com.baidu.hugegraph.metrics.MetricsUtil;
+import com.baidu.hugegraph.metrics.ServerReporter;
 import com.baidu.hugegraph.serializer.JsonSerializer;
 import com.baidu.hugegraph.serializer.Serializer;
 import com.baidu.hugegraph.server.RestServer;
